Fix post-sync Release build errors

+++ source/src/Runtime/MultiplayerClientProbe.cs
@@ -580,7 +580,7 @@ internal static class MultiplayerClientProbe
                 break;
             case MadScience madScience:
                 fingerprint.Add((int)madScience.TinkerTimeType);
-                fingerprint.Add(madScience.TinkerTimeRider);
+                fingerprint.Add((int)madScience.TinkerTimeRider);
                 break;
             case Rampage rampage:
                 fingerprint.Add(rampage.ExtraDamageFromPlays);
